backend/requirements.txt
openai-agents
openai
python-dotenv
fastapi
pydantic
pydantic-settings
uvicorn
pytest
requests
pytest-cov
httpx
pytest-mock
google-adk
litellm
beautifulsoup4
